requirements.txt
# Python dependencies for CryptoPulse

# Core Spark and Delta Lake
pyspark==3.5.3
delta-spark==3.2.0

# Data processing
pandas==2.1.4
numpy==1.26.2

# API and data fetching
requests==2.31.0
ccxt==4.1.92

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
